Conflicts:
	Lib/NativeBinaries/x86/git2.dll
	Lib/NativeBinaries/x86/git2.pdb
	LibGit2Sharp.Tests/BranchFixture.cs
	LibGit2Sharp.Tests/CommitFixture.cs
	LibGit2Sharp.Tests/ReferenceFixture.cs
	LibGit2Sharp/Configuration.cs
	LibGit2Sharp/Core/Ensure.cs
	LibGit2Sharp/Core/GitError.cs
	LibGit2Sharp/RemoteCollection.cs

+++ LibGit2Sharp/Core/UnSafeNativeMethods.cs
@@ -28,7 +28,7 @@ namespace LibGit2Sharp.Core
         internal struct git_strarray
         {
             public sbyte** strings;
-            public IntPtr size;
+            public uint size;
         }
 
         #endregion
